Added Position/Scale Transform

+++ CG_Engine/include/CG/Color.hpp
@@ -4,7 +4,7 @@ namespace CG {
 
 struct Color
 {
-	double r, g, b, a;
+	float r, g, b, a;
 };
 
 }
